+++ lbr_normal_analysis_roughness/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.8)
-project(lbr_terrain_analysis)
+project(lbr_normal_analysis_roughness)
 
 if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   add_compile_options(-Wall -Wextra -Wpedantic)
@@ -13,12 +13,12 @@ find_package(PCL REQUIRED COMPONENTS filters features)
 
 ament_auto_find_build_dependencies()
 
-ament_auto_add_library(lbr_terrain_analysis src/lbr_terrain_analysis.cpp)
-rclcpp_components_register_nodes(lbr_terrain_analysis "lbr_terrain_analysis::TerrainAnalysis")
+ament_auto_add_library(lbr_normal_analysis_roughness src/lbr_normal_analysis_roughness.cpp)
+rclcpp_components_register_nodes(lbr_normal_analysis_roughness "lbr_normal_analysis_roughness::TerrainAnalysis")
 
-ament_auto_add_executable(lbr_terrain_analysis_node
-  src/lbr_terrain_analysis_node.cpp
-  src/lbr_terrain_analysis.cpp)
+ament_auto_add_executable(lbr_normal_analysis_roughness_node
+  src/lbr_normal_analysis_roughness_node.cpp
+  src/lbr_normal_analysis_roughness.cpp)
 
 # install launch file
 install(DIRECTORY
@@ -30,12 +30,12 @@ install(DIRECTORY
 # gtest
 if(BUILD_TESTING)
   find_package(ament_cmake_gtest REQUIRED)
-  ament_add_gtest(unit_test_lbr_terrain_analysis test/unit_test_lbr_terrain_analysis.cpp)
-  target_include_directories(unit_test_lbr_terrain_analysis PUBLIC
+  ament_add_gtest(unit_test_lbr_normal_analysis_roughness test/unit_test_lbr_normal_analysis_roughness.cpp)
+  target_include_directories(unit_test_lbr_normal_analysis_roughness PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
     $<INSTALL_INTERFACE:include>
   )
-  target_link_libraries(unit_test_lbr_terrain_analysis lbr_terrain_analysis)
+  target_link_libraries(unit_test_lbr_normal_analysis_roughness lbr_normal_analysis_roughness)
 endif()
 
 if(BUILD_TESTING)
