+++ lbr_normal_analysis_roughness/config/lbr_normal_analysis_roughness.rviz
