+++ lbr_normal_analysis_roughness/include/lbr_normal_analysis_roughness/visibility_control.h
@@ -1,0 +1,49 @@
+// Copyright (c) 2024 Tohoku Univ. Space Robotics Lab.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef lbr_normal_analysis_roughness__VISIBILITY_CONTROL_H_
+#define lbr_normal_analysis_roughness__VISIBILITY_CONTROL_H_
+
+// This logic was borrowed (then namespaced) from the examples on the gcc wiki:
+//     https://gcc.gnu.org/wiki/Visibility
+
+#if defined _WIN32 || defined __CYGWIN__
+#ifdef __GNUC__
+#define lbr_normal_analysis_roughness_EXPORT __attribute__((dllexport))
+#define lbr_normal_analysis_roughness_IMPORT __attribute__((dllimport))
+#else
+#define lbr_normal_analysis_roughness_EXPORT __declspec(dllexport)
+#define lbr_normal_analysis_roughness_IMPORT __declspec(dllimport)
+#endif
+#ifdef lbr_normal_analysis_roughness_BUILDING_LIBRARY
+#define lbr_normal_analysis_roughness_PUBLIC lbr_normal_analysis_roughness_EXPORT
+#else
+#define lbr_normal_analysis_roughness_PUBLIC lbr_normal_analysis_roughness_IMPORT
+#endif
+#define lbr_normal_analysis_roughness_PUBLIC_TYPE lbr_normal_analysis_roughness_PUBLIC
+#define lbr_normal_analysis_roughness_LOCAL
+#else
+#define lbr_normal_analysis_roughness_EXPORT __attribute__((visibility("default")))
+#define lbr_normal_analysis_roughness_IMPORT
+#if __GNUC__ >= 4
+#define lbr_normal_analysis_roughness_PUBLIC __attribute__((visibility("default")))
+#define lbr_normal_analysis_roughness_LOCAL __attribute__((visibility("hidden")))
+#else
+#define lbr_normal_analysis_roughness_PUBLIC
+#define lbr_normal_analysis_roughness_LOCAL
+#endif
+#define lbr_normal_analysis_roughness_PUBLIC_TYPE
+#endif
+
+#endif  // lbr_normal_analysis_roughness__VISIBILITY_CONTROL_H_
