+++ lbr_normal_analysis_roughness/test/unit_test_lbr_normal_analysis_roughness.cpp
@@ -12,18 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "lbr_terrain_analysis/lbr_terrain_analysis.hpp"
+#include <gtest/gtest.h>
 
-int main(int argc, char * argv[])
+#include "lbr_normal_analysis_roughness/lbr_normal_analysis_roughness.hpp"
+
+namespace lbr_normal_analysis_roughness
 {
-  rclcpp::init(argc, argv);
-  rclcpp::executors::MultiThreadedExecutor exec;
 
-  const auto lbr_terrain_analysis =
-    std::make_shared<lbr_terrain_analysis::TerrainAnalysis>(rclcpp::NodeOptions());
-  exec.add_node(lbr_terrain_analysis);
-
-  exec.spin();
-
-  rclcpp::shutdown();
+TEST(LbrTerrainAnalysis, oneTestCase)
+{
 }
+
+int main(int argc, char ** argv)
+{
+  testing::InitGoogleTest(&argc, argv);
+  return RUN_ALL_TESTS();
+}
+
+}  // namespace lbr_normal_analysis_roughness
